fix(person): Yes/No fields
Fix the bug on the yes/no fields.

The problem was that the checkbox returned a null value when checked. Now they return 1 if checked. They aren't in the request id not checked.

When edit, I also had to compare the difference between the stored form and the request received. All fields not present in the request are detached.

issue: #293
Signed-off-by: Lduf <lucas.dufour@insa-lyon.fr>